requirements.txt
python-dotenv~=1.0.1
tavily-python==0.3.3
duckduckgo_search==5.3.1b1 
beautifulsoup4==4.12.3
openai==1.82.0
langgraph==0.3.21
langchain_core==0.3.62
langchain_openai==0.3.18
langchain-community==0.3.24
langchain_experimental==0.3.4
langchain-chroma==0.2.4
langchain==0.3.25
aiosqlite==0.20.0
pygments==2.17.2
pygraphviz==1.13
gradio==4.31.3
httpx==0.27.2
pypdf==5.4.0
qdrant-client
langchain-google-genai==2.1.5
grandalf==0.8
schema==0.7.7
langgraph-checkpoint-sqlite==2.0.10
